# Dimension dictionary reduce locking