--- a/processing/src/main/java/org/apache/druid/segment/DimensionDictionary.java
+++ b/processing/src/main/java/org/apache/druid/segment/DimensionDictionary.java
@@ -27,7 +27,7 @@ import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
+import java.util.concurrent.locks.StampedLock;
 
 /**
  * Buildable dictionary for some comparable type. Values are unsorted, or rather sorted in the order which they are
@@ -51,18 +51,18 @@ public abstract class DimensionDictionary<T extends Comparable<T>>
   private final Object2IntMap<T> valueToId = new Object2IntOpenHashMap<>();
 
   private final List<T> idToValue = new ArrayList<>();
-  private final ReentrantReadWriteLock lock;
+  private final StampedLock lock;
 
   public DimensionDictionary(Class<T> cls)
   {
     this.cls = cls;
-    this.lock = new ReentrantReadWriteLock();
+    this.lock = new StampedLock();
     valueToId.defaultReturnValue(ABSENT_VALUE_ID);
   }
 
   public int getId(@Nullable T value)
   {
-    lock.readLock().lock();
+    long stamp = lock.readLock();
     try {
       if (value == null) {
         return idForNull;
@@ -70,14 +70,14 @@ public abstract class DimensionDictionary<T extends Comparable<T>>
       return valueToId.getInt(value);
     }
     finally {
-      lock.readLock().unlock();
+      lock.unlockRead(stamp);
     }
   }
 
   @Nullable
   public T getValue(int id)
   {
-    lock.readLock().lock();
+    long stamp = lock.readLock();
     try {
       if (id == idForNull) {
         return null;
@@ -85,7 +85,7 @@ public abstract class DimensionDictionary<T extends Comparable<T>>
       return idToValue.get(id);
     }
     finally {
-      lock.readLock().unlock();
+      lock.unlockRead(stamp);
     }
   }
 
@@ -93,7 +93,7 @@ public abstract class DimensionDictionary<T extends Comparable<T>>
   {
     T[] values = (T[]) Array.newInstance(cls, ids.length);
 
-    lock.readLock().lock();
+    long stamp = lock.readLock();
     try {
       for (int i = 0; i < ids.length; i++) {
         values[i] = (ids[i] == idForNull) ? null : idToValue.get(ids[i]);
@@ -101,19 +101,19 @@ public abstract class DimensionDictionary<T extends Comparable<T>>
       return values;
     }
     finally {
-      lock.readLock().unlock();
+      lock.unlockRead(stamp);
     }
   }
 
   public int size()
   {
-    lock.readLock().lock();
+    long stamp = lock.readLock();
     try {
       // using idToValue rather than valueToId because the valueToId doesn't account null value, if it is present.
       return idToValue.size();
     }
     finally {
-      lock.readLock().unlock();
+      lock.unlockRead(stamp);
     }
   }
 
@@ -133,7 +133,7 @@ public abstract class DimensionDictionary<T extends Comparable<T>>
 
   public int add(@Nullable T originalValue)
   {
-    lock.writeLock().lock();
+    long stamp = lock.writeLock();
     try {
       if (originalValue == null) {
         if (idForNull == ABSENT_VALUE_ID) {
@@ -160,29 +160,29 @@ public abstract class DimensionDictionary<T extends Comparable<T>>
       return index;
     }
     finally {
-      lock.writeLock().unlock();
+      lock.unlockWrite(stamp);
     }
   }
 
   public T getMinValue()
   {
-    lock.readLock().lock();
+    long stamp = lock.readLock();
     try {
       return minValue;
     }
     finally {
-      lock.readLock().unlock();
+      lock.unlockRead(stamp);
     }
   }
 
   public T getMaxValue()
   {
-    lock.readLock().lock();
+    long stamp = lock.readLock();
     try {
       return maxValue;
     }
     finally {
-      lock.readLock().unlock();
+      lock.unlockRead(stamp);
     }
   }
 
@@ -193,12 +193,12 @@ public abstract class DimensionDictionary<T extends Comparable<T>>
 
   public SortedDimensionDictionary<T> sort()
   {
-    lock.readLock().lock();
+    long stamp = lock.readLock();
     try {
       return new SortedDimensionDictionary<>(idToValue, idToValue.size());
     }
     finally {
-      lock.readLock().unlock();
+      lock.unlockRead(stamp);
     }
   }