### Description

Following from https://github.com/apache/druid/pull/12105 with 1-20% CPU in `DimensionDictionary`, there are some improvements to be made to the `add()` path on the `StringDimensionIndexer`. The main concern is that the locking is more complicated than necessary.

There are three sub-changes:
- Introduce a benchmark
- Swap the existing lock implementation to use a StampedLock
- Reduce the number of times we have to cross the lock

Estimate this code reduces CPU by ~5-7% on our indexing jobs overall, with ~30% improvements to the uncontended path.

### Related

#12105, #12109, 

##### Key changed/added classes in this PR
 * Introduce benchmark `StringDimensionIndexerProcessBenchmark`
 * `DimensionDictionary` locking changes

<hr>

This PR has:

- [x] been self-reviewed.
   - [x] using the [concurrency checklist](https://github.com/apache/druid/blob/master/dev/code-review/concurrency.md) (Remove this item if the PR doesn't have any relation to concurrency.)
- [x] been tested in a test Druid cluster. -- a solution more like 12105 on top of 0.20 is currently running in Production for many months.