@clintropolis 2023-01-26
it would probably be worth checking that query performance for realtime tasks is not negatively impacted too much by this change (especially from multiple concurrent threads). There is only one writer thread but potentially many reader threads, as new dimension selectors are created on top of this dictionary during realtime queries.

For batch ingestion tasks, it might be worth using a separate `DimensionDictionary` implementation that has no locks. I use a similar thing for the `NestedDataColumnIndexer`, `ComparatorDimensionDictionary`, which is basically a copy of `DimensionDictionary` except uses a `Comparator` instead of requiring a `Comparable`, and doesn't have any locks since the nested column indexer doesn't currently use the dictionary it is building for realtime queries.