@clintropolis 2023-01-28
cool, thanks for checking. Btw, I actually had a bug in that lockless dictionary used by nested columns that resulted in it not fully merging dictionaries which is fixed in this PR https://github.com/apache/druid/pull/13714/files#diff-b00489e26a5bff0e308349f4bdf8dc21a55798f8d0f35c995c56599cb6485eb1R93 😅

anyway, I'll try to have a closer look at this PR sometime soon